requirements.txt
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
stripe==2.21.0
tensorflow==2.0.0
pillow==6.2.1
fastai==1.0.57